+++ api/source/utils/config.js
@@ -40,7 +40,7 @@ const config = {
         directory: process.env.STIGMAN_CLIENT_DIRECTORY || '../../client/dist',
         extraScopes: process.env.STIGMAN_CLIENT_EXTRA_SCOPES,
         scopePrefix: process.env.STIGMAN_CLIENT_SCOPE_PREFIX,
-        responseMode: process.env.STIGMAN_CLIENT_RESPONSE_MODE || "fragment",
+        responseMode: process.env.STIGMAN_CLIENT_RESPONSE_MODE || "query",
         reauthAction: process.env.STIGMAN_CLIENT_REAUTH_ACTION || "popup",
         strictPkce: process.env.STIGMAN_CLIENT_STRICT_PKCE !== 'false',
         stateEvents: process.env.STIGMAN_CLIENT_STATE_EVENTS !== 'false',